- mdmkolbe fixed build errors in the new event system and a typo   that used button instead of buttonState.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@23202 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csutil/event.h
@@ -98,11 +98,11 @@ public:
     uint32 AxesChanged, uint button, bool buttonstate, uint32 buttonMask, 
     uint32 modifiers);
   static csEvent *NewEvent (csRef<iEventNameRegistry> &reg, csTicks, 
-    csEventID name, int n, csMouseEventType etype, int x, int y, 
+    csEventID name, uint8 n, csMouseEventType etype, int x, int y, 
     uint32 axesChanged, uint button, bool buttonstate, uint32 buttonMask, 
     uint32 modifiers);
   static csEvent *NewEvent (csRef<iEventNameRegistry> &reg, csTicks, 
-    csEventID name, int n, csMouseEventType etype, const int32 *axes, 
+    csEventID name, uint8 n, csMouseEventType etype, const int32 *axes, 
     uint8 numAxes, uint32 axesChanged, uint button, bool buttonstate, 
     uint32 buttonMask, uint32 modifiers);
   //@}
